Separate diagnostics tests into their own class
The only change to the logic is that the tests now use RemoteInvoke to run each test in its own process, so as to avoid EventSource and DiagnosticSource interference.

Also removed the xunit attributes from the assembly that prevented parallelization.  It's no longer needed as of this change.
