+++ src/System.Net.Http/tests/FunctionalTests/HttpClientTest.cs
@@ -393,153 +393,6 @@ namespace System.Net.Http.Functional.Tests
             }
         }
 
-        /// <remarks>
-        /// This test must be in the same test collection as any others testing HttpClient/WinHttpHandler
-        /// DiagnosticSources, since the global logging mechanism makes them conflict inherently.
-        /// </remarks>
-        [Fact]
-        public void SendAsync_ExpectedDiagnosticSourceLogging()
-        {
-            bool requestLogged = false;
-            Guid requestGuid = Guid.Empty;
-            bool responseLogged = false;
-            Guid responseGuid = Guid.Empty;
-
-            var diagnosticListenerObserver = new FakeDiagnosticListenerObserver(
-                kvp =>
-                {
-                    if (kvp.Key.Equals("System.Net.Http.Request"))
-                    {
-                        Assert.NotNull(kvp.Value);
-                        GetPropertyValueFromAnonymousTypeInstance<HttpRequestMessage>(kvp.Value, "Request");
-                        requestGuid = GetPropertyValueFromAnonymousTypeInstance<Guid>(kvp.Value, "LoggingRequestId");
-
-                        requestLogged = true;
-                    }
-                    else if (kvp.Key.Equals("System.Net.Http.Response"))
-                    {
-                        Assert.NotNull(kvp.Value);
-
-                        GetPropertyValueFromAnonymousTypeInstance<HttpResponseMessage>(kvp.Value, "Response");
-                        responseGuid = GetPropertyValueFromAnonymousTypeInstance<Guid>(kvp.Value, "LoggingRequestId");
-
-                        responseLogged = true;
-                    }
-                });
-
-            using (DiagnosticListener.AllListeners.Subscribe(diagnosticListenerObserver))
-            {
-                diagnosticListenerObserver.Enable();
-                using (var client = new HttpClient())
-                {
-                    var response = client.GetAsync(HttpTestServers.RemoteEchoServer).Result;
-                }
-
-                Assert.True(requestLogged, "Request was not logged.");
-                // Poll with a timeout since logging response is not synchronized with returning a response.
-                WaitForTrue(() => responseLogged, TimeSpan.FromSeconds(1), "Response was not logged within 1 second timeout.");
-                Assert.Equal(requestGuid, responseGuid);
-                diagnosticListenerObserver.Disable();
-            }
-        }
-
-        /// <remarks>
-        /// This test must be in the same test collection as any others testing HttpClient/WinHttpHandler
-        /// DiagnosticSources, since the global logging mechanism makes them conflict inherently.
-        /// </remarks>
-        [Fact]
-        public void SendAsync_ExpectedDiagnosticSourceNoLogging()
-        {
-            bool requestLogged = false;
-            bool responseLogged = false;
-
-            var diagnosticListenerObserver = new FakeDiagnosticListenerObserver(
-                kvp =>
-                {
-                    if (kvp.Key.Equals("System.Net.Http.Request"))
-                    {
-                        requestLogged = true;
-                    }
-                    else if (kvp.Key.Equals("System.Net.Http.Response"))
-                    {
-                        responseLogged = true;
-                    }
-                });
-
-            using (DiagnosticListener.AllListeners.Subscribe(diagnosticListenerObserver))
-            {
-                using (var client = new HttpClient())
-                {
-                    var response = client.GetAsync(HttpTestServers.RemoteEchoServer).Result;
-                }
-
-                Assert.False(requestLogged, "Request was logged while logging disabled.");
-                WaitForFalse(() => responseLogged, TimeSpan.FromSeconds(1), "Response was logged while logging disabled.");
-            }
-        }
-
-        [Fact]
-        public async Task SendAsync_HttpTracingEnabled_Succeeds()
-        {
-            using (var listener = new TestEventListener("Microsoft-System-Net-Http", EventLevel.Verbose))
-            {
-                var events = new ConcurrentQueue<EventWrittenEventArgs>();
-                await listener.RunWithCallbackAsync(events.Enqueue, async () =>
-                {
-                    // Exercise various code paths to get coverage of tracing
-                    using (var client = new HttpClient())
-                    {
-                        // Do a get to a loopback server
-                        await LoopbackServer.CreateServerAsync(async (server, url) =>
-                        {
-                            await TestHelper.WhenAllCompletedOrAnyFailed(
-                                LoopbackServer.ReadRequestAndSendResponseAsync(server),
-                                client.GetAsync(url));
-                        });
-
-                        // Do a post to a remote server
-                        byte[] expectedData = Enumerable.Range(0, 20000).Select(i => (byte)i).ToArray();
-                        HttpContent content = new ByteArrayContent(expectedData);
-                        content.Headers.ContentMD5 = TestHelper.ComputeMD5Hash(expectedData);
-                        using (HttpResponseMessage response = await client.PostAsync(HttpTestServers.RemoteEchoServer, content))
-                        {
-                            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-                        }
-                    }
-                });
-
-                // We don't validate receiving specific events, but rather that we do at least
-                // receive some events, and that enabling tracing doesn't cause other failures
-                // in processing.
-                Assert.InRange(events.Count, 1, int.MaxValue);
-            }
-        }
-
-        private void WaitForTrue(Func<bool> p, TimeSpan timeout, string message)
-        {
-            // Assert that spin doesn't time out.
-            Assert.True(System.Threading.SpinWait.SpinUntil(p, timeout), message);
-        }
-
-        private void WaitForFalse(Func<bool> p, TimeSpan timeout, string message)
-        {
-            // Assert that spin times out.
-            Assert.False(System.Threading.SpinWait.SpinUntil(p, timeout), message);
-        }
-
-        private T GetPropertyValueFromAnonymousTypeInstance<T>(object obj, string propertyName)
-        {
-            Type t = obj.GetType();
-
-            PropertyInfo p = t.GetRuntimeProperty(propertyName);
-
-            object propertyValue = p.GetValue(obj);
-            Assert.NotNull(propertyValue);
-            Assert.IsAssignableFrom<T>(propertyValue);
-
-            return (T)propertyValue;
-        }
-
         private static string CreateFakeUri() => $"http://{Guid.NewGuid().ToString("N")}";
 
         private static async Task<T> WhenCanceled<T>(CancellationToken cancellationToken)
